NFC: Fix for HCI network detection issue
The HCI network is not detected anymore after a FW upgrade as the
wrong status is delivered to the calling layer.

The pointer p_msg is only dereferenced in certain cases.
The NULL pointer check should therefore only be issued if
the pointer is actually dereferenced.

Modified the NULL check to check the correct case.

Change-Id: I5e25dc5bfbb8823cde26fe6380cf2e77fae00d2f
CRs-Fixed: 990176